Move AndroidTests to Android Project.  All tests passing
