+++ android/CouchbaseLite/src/androidTest/java/com/couchbase/lite/internal/utils/IOUtils.java
@@ -15,7 +15,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //
-package com.couchbase.lite.utils;
+package com.couchbase.lite.internal.utils;
 
 import java.io.ByteArrayOutputStream;
 import java.io.File;
