+++ android/CouchbaseLite/src/androidTest/java/com/couchbase/lite/internal/utils/StringUtilsTest.java
@@ -21,6 +21,7 @@ import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
 
+
 public class StringUtilsTest {
     @Test
     public void testStringByDeletingLastPathComponent() {
