--- shared/src/androidTest/java/com/couchbase/litecore/C4NestedQueryTest.java
@@ -1,55 +1,0 @@
-//
-// C4NestedQueryTest.java
-//
-// Copyright (c) 2017 Couchbase, Inc All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-package com.couchbase.litecore;
-
-import org.junit.Test;
-
-import java.util.Arrays;
-
-import static org.junit.Assert.assertEquals;
-
-public class C4NestedQueryTest extends C4QueryBaseTest {
-
-    C4Query query = null;
-
-    //-------------------------------------------------------------------------
-    // public methods
-    //-------------------------------------------------------------------------
-
-    @Override
-    public void setUp() throws Exception {
-        super.setUp();
-        importJSONLines("nested.json");
-    }
-
-    @Override
-    public void tearDown() throws Exception {
-        if (query != null) {
-            query.free();
-            query = null;
-        }
-        super.tearDown();
-    }
-
-    // - DB Query ANY nested
-    @Test
-    public void testDBQueryANYNested() throws LiteCoreException, Exception {
-        compile(json5("['ANY', 'Shape', ['.', 'shapes'], ['=', ['?', 'Shape', 'color'], 'red']]"));
-        assertEquals(Arrays.asList("0000001", "0000003"), run());
-    }
-}
