--- shared/src/androidTest/java/com/couchbase/litecore/utils/Utils.java
@@ -1,94 +1,0 @@
-//
-// Utils.java
-//
-// Copyright (c) 2017 Couchbase, Inc All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-package com.couchbase.litecore.utils;
-
-import android.util.Log;
-
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.FilenameFilter;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-
-public class Utils {
-    public static boolean deleteRecursive(File fileOrDirectory) {
-        if (fileOrDirectory.isDirectory()) {
-            for (File child : fileOrDirectory.listFiles()) {
-                deleteRecursive(child);
-            }
-        }
-        return fileOrDirectory.delete() || !fileOrDirectory.exists();
-    }
-
-    public static boolean deleteFiles(final File directory, final String prefix) {
-        for (File child : directory.listFiles(new FilenameFilter() {
-            @Override
-            public boolean accept(File dir, String name) {
-                return name.startsWith(prefix);
-            }
-        })) {
-            deleteRecursive(child);
-        }
-        return true;
-    }
-
-    public static void copyAssets(String filename, File outDir) {
-        InputStream in = null;
-        OutputStream out = null;
-        try {
-            in = Utils.class.getResourceAsStream("/assets/" + filename);
-            if (!outDir.exists())
-                outDir.mkdirs();
-            File outFile = new File(outDir, filename);
-            out = new FileOutputStream(outFile);
-            copyFile(in, out);
-        } catch (IOException e) {
-            Log.e("tag", "Failed to copy asset file: " + filename, e);
-        } finally {
-            if (in != null) {
-                try {
-                    in.close();
-                } catch (IOException e) {
-                    // NOOP
-                }
-            }
-            if (out != null) {
-                try {
-                    out.close();
-                } catch (IOException e) {
-                    // NOOP
-                }
-            }
-        }
-    }
-
-    public static void copyFile(InputStream in, OutputStream out) throws IOException {
-        byte[] buffer = new byte[1024];
-        int read;
-        while ((read = in.read(buffer)) != -1) {
-            out.write(buffer, 0, read);
-        }
-    }
-
-    public static String readFile(String path) throws IOException {
-        return new String(Files.readAllBytes(Paths.get(path)));
-    }
-}
