--- shared/src/main/java/com/couchbase/lite/internal/utils/FileUtils.java
@@ -1,38 +1,0 @@
-//
-// FileUtils.java
-//
-// Copyright (c) 2017 Couchbase, Inc All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-package com.couchbase.lite.internal.utils;
-
-import java.io.File;
-
-public final class FileUtils {
-    private FileUtils() {
-    }
-
-    public static boolean deleteRecursive(String fileOrDirectory) {
-        return deleteRecursive(new File(fileOrDirectory));
-    }
-
-    public static boolean deleteRecursive(File fileOrDirectory) {
-        if (fileOrDirectory.isDirectory()) {
-            for (File child : fileOrDirectory.listFiles()) {
-                deleteRecursive(child);
-            }
-        }
-        return fileOrDirectory.delete() || !fileOrDirectory.exists();
-    }
-}
